GeneCraft tourne maintenant sur la version 2.78 de Bullet !
+ Contraintes destructives
* Correction du bug des rotationnals motors (nouveau type de genomes!)
* Réaorganisation dans l'achitecture des factories + world
* Toujours un peu d'interface
* Correction d'un bug de type des sorties du cerveau (les positions des connexions étaient en int...)



git-svn-id: http://www.genecraft-project.com/svn/genecraft/trunk@135 2ce2bb41-2ff2-4fd5-96a4-e3c75a7bd90e

+++ src/genelab/core/entities/brain/brainfunctional.cpp
@@ -80,7 +80,7 @@ float sigmoid(float x)
                 tree.append(new BrainNode(t));
                 break;
             case IN:
-                tree.append(new BrainNodeIn(nodePart[1].toInt(), nodePart[2].toInt()));
+                tree.append(new BrainNodeIn(nodePart[1].toFloat(), nodePart[2].toFloat()));
                 break;
             case CONST:
                 tree.append(new BrainNodeConst(nodePart[1].toFloat()));
